+++ rust/src/nasl/builtin/ssh/mod.rs
@@ -1,14 +1,16 @@
 #[cfg(feature = "nasl-builtin-libssh")]
 mod libssh;
 #[cfg(feature = "nasl-builtin-libssh")]
-pub use libssh::Ssh;
+pub use libssh::{SessionId, Socket, Ssh, SshError};
 
 #[cfg(not(feature = "nasl-builtin-libssh"))]
 mod russh;
 #[cfg(not(feature = "nasl-builtin-libssh"))]
-pub use russh::Ssh;
+pub use russh::{SessionId, Socket, Ssh, SshError};
 
 #[cfg(test)]
 mod tests;
 
 mod utils;
+
+mod impls;
